refactoring of dataownerapi
