+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/controller/CompanyDataController.kt
@@ -7,6 +7,7 @@ import org.dataland.datalandbackend.model.DataType
 import org.dataland.datalandbackend.model.StoredCompany
 import org.dataland.datalandbackend.model.companies.AggregatedFrameworkDataSummary
 import org.dataland.datalandbackend.model.companies.CompanyAvailableDistinctValues
+import org.dataland.datalandbackend.model.companies.CompanyDataOwners
 import org.dataland.datalandbackend.model.companies.CompanyId
 import org.dataland.datalandbackend.model.companies.CompanyInformation
 import org.dataland.datalandbackend.model.companies.CompanyInformationPatch
@@ -15,6 +16,7 @@ import org.dataland.datalandbackend.repositories.CompanyIdentifierRepository
 import org.dataland.datalandbackend.repositories.utils.StoredCompanySearchFilter
 import org.dataland.datalandbackend.services.CompanyAlterationManager
 import org.dataland.datalandbackend.services.CompanyQueryManager
+import org.dataland.datalandbackend.services.DataOwnersManager
 import org.dataland.datalandbackendutils.exceptions.ResourceNotFoundApiException
 import org.dataland.keycloakAdapter.auth.DatalandAuthentication
 import org.slf4j.LoggerFactory
@@ -22,6 +24,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.ResponseEntity
 import org.springframework.orm.jpa.JpaObjectRetrievalFailureException
 import org.springframework.web.bind.annotation.RestController
+import java.util.UUID
 
 /**
  * Controller for the company data endpoints
@@ -35,6 +38,7 @@ class CompanyDataController(
     @Autowired private val companyAlterationManager: CompanyAlterationManager,
     @Autowired private val companyQueryManager: CompanyQueryManager,
     @Autowired private val companyIdentifierRepositoryInterface: CompanyIdentifierRepository,
+    @Autowired private val dataOwnersManager: DataOwnersManager,
 ) : CompanyApi {
     private val logger = LoggerFactory.getLogger(javaClass)
 
@@ -176,5 +180,40 @@ class CompanyDataController(
             companyQueryManager
                 .getCompanyApiModelById(companyId, DatalandAuthentication.fromContextOrNull()).companyInformation,
         )
+    }
+
+    override fun postDataOwner(companyId: UUID, userId: UUID): ResponseEntity<CompanyDataOwners> {
+        logger.info("Received a request to post a data owner with Id $userId to company with Id $companyId.")
+        val companyDataOwnersEntity = dataOwnersManager.addDataOwnerToCompany(companyId.toString(), userId.toString())
+        return ResponseEntity.ok(
+            CompanyDataOwners(
+                companyId = companyDataOwnersEntity.companyId,
+                dataOwners = companyDataOwnersEntity.dataOwners,
+            ),
+        )
+    }
+    override fun getDataOwners(companyId: UUID): ResponseEntity<List<String>> {
+        logger.info("Received a request to get a data owner from company Id $companyId.")
+        val companyDataOwnersEntity = dataOwnersManager.getDataOwnerFromCompany(companyId.toString())
+        return ResponseEntity.ok(companyDataOwnersEntity.dataOwners)
+    }
+
+    override fun deleteDataOwner(companyId: UUID, userId: UUID): ResponseEntity<CompanyDataOwners> {
+        logger.info("Received a request to delete a data owner with Id $userId to company with Id $companyId.")
+        val companyDataOwnersEntity = dataOwnersManager.deleteDataOwnerFromCompany(
+            companyId.toString(),
+            userId.toString(),
+        )
+        return ResponseEntity.ok(
+            CompanyDataOwners(
+                companyId = companyDataOwnersEntity.companyId,
+                dataOwners = companyDataOwnersEntity.dataOwners,
+            ),
+        )
+    }
+
+    override fun isUserDataOwnerForCompany(companyId: UUID, userId: UUID) {
+        logger.info("Received a request to check if user with Id $userId is data owner of company with Id $companyId.")
+        dataOwnersManager.checkUserCompanyCombinationForDataOwnership(companyId.toString(), userId.toString())
     }
 }
